Move getPattern so it can be used at runtime
