+++ packages/astro/src/core/routing/manifest/pattern.ts
@@ -1,0 +1,57 @@
+import type {
+	AstroConfig,
+	RoutePart,
+} from '../../../@types/astro.js';
+
+export function getPattern(
+	segments: RoutePart[][],
+	base: AstroConfig['base'],
+	addTrailingSlash: AstroConfig['trailingSlash']
+) {
+	const pathname = segments
+		.map((segment) => {
+			if (segment.length === 1 && segment[0].spread) {
+				return '(?:\\/(.*?))?';
+			} else {
+				return (
+					'\\/' +
+					segment
+						.map((part) => {
+							if (part.spread) {
+								return '(.*?)';
+							} else if (part.dynamic) {
+								return '([^/]+?)';
+							} else {
+								return part.content
+									.normalize()
+									.replace(/\?/g, '%3F')
+									.replace(/#/g, '%23')
+									.replace(/%5B/g, '[')
+									.replace(/%5D/g, ']')
+									.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+							}
+						})
+						.join('')
+				);
+			}
+		})
+		.join('');
+
+	const trailing =
+		addTrailingSlash && segments.length ? getTrailingSlashPattern(addTrailingSlash) : '$';
+	let initial = '\\/';
+	if (addTrailingSlash === 'never' && base !== '/') {
+		initial = '';
+	}
+	return new RegExp(`^${pathname || initial}${trailing}`);
+}
+
+function getTrailingSlashPattern(addTrailingSlash: AstroConfig['trailingSlash']): string {
+	if (addTrailingSlash === 'always') {
+		return '\\/$';
+	}
+	if (addTrailingSlash === 'never') {
+		return '$';
+	}
+	return '\\/?$';
+}
